internal/instance: Add GetOSType
Seems on VMware's end, all Windows OSes are prefixed with 'win', so
that's the heuristic used here.

This is a bit brittle as it assumes a Linux VM will never have a version
prefixed with 'win', but the alternative is to have to maintain a list
of Windows versions into the future.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>