cmd/migration-managerd/internal/api: Refactor source syncing
Splits up the helpers and changes the logic to sync only one source
given as argument.

These helpers no longer skip errors by default. If the caller doesn't
care about errors then it should log and skip them. This means if an
error is encountered in a transaction, the transaction will not be
committed.

Some efficiency improvements by fetching all records from the database
earlier on and storing them in a map rather than n*m queries for each
instance in each source.

Signed-off-by: Max Asnaashari <max.asna@futurfusion.io>