cmd/migration-managerd/internal/api: Refactor processQueuedBatches
This one received the most substantial change.

There is now no longer a bunch of loose goroutines spawned at the end of
this function.

Instead, the flow works like this:
Every 10s, in processQueuedBatches:
 * Get all batches
 * Filter only the batches we care about (QUEUED, in window)
 * Filter only the instances we care about (ASSIGNED)
 * Set batches to RUNNING, and their instances to CREATING

Every 10s, in initMigrationWorkers:
 * Get all RUNNING batches and their CREATING instances
 * Concurrently:
   * Create VMs on each instance
   * Exec the migration worker for each instance
   * If an error occurs, set the instance to ERRORED
   * Optionally delete the target VM

This way, we still have the benefits of the goroutines, but they are no
longer dangling, and risk interfering with each other or the next
periodic task.

By filtering on state, we also avoid processQueuedBatches clashing with
other runs of initMigrationWorkers or finalizeCompleteInstances.

Signed-off-by: Max Asnaashari <max.asna@futurfusion.io>