cmd/migration-managerd/internal/api: Refactor finalizeCompleteInstances
Like the others helpers, this is changed to properly handle errors and
use a reverter to set all instances to MIGRATIONSTATUS_ERROR if an error
occurs at any point, as well as splits out target/source connections
from the db transaction.

Configuration for each instance is split off into its own helper so if
any one instance fails, the others will continue to be reconfigured.

Signed-off-by: Max Asnaashari <max.asna@futurfusion.io>