internal/transaction: Add helper to extract TX from DBTX
generate-database uses Stmt(*sql.Stmt) whose implementation will ignore
a running transaction if passed a `dbtx`.

Since we can't know if dbtx is a transaction from the generator,
and the generator should not dictate how dbtx may wrap a transaction
(via the context like we do or some other mechanism),
it is up to the caller to pass in the correct type.

Signed-off-by: Max Asnaashari <max.asna@futurfusion.io>